fix typo in csv records

+++ src-tauri/src/csvparsing/records.rs
@@ -33,10 +33,10 @@ pub struct SellerRecord {
 #[derive(Debug, Deserialize, Serialize)]
 pub struct InvoiceRecord {
     id: i64,
-    pub total: i64,
+    pub total: f32,
     pub status: String,
     pub created_at: String,
-    pub client_id: i64,
+    pub client_id: i32,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
